moved components and pages folders to src folder
