+++ src/App.css
@@ -36,3 +36,12 @@
     transform: rotate(360deg);
   }
 }
+
+.container {
+  max-width: 1920px;
+}
+.navbar-container {
+  display: fixed;
+  height: 50px;
+  width: 100%;
+}
