+++ src/pages/details-info.jsx
